Update provider resources code and go.mod.
Since the provider is located at "github.com/ansible/terraform-provider-ansible",
repository links in go.mod and main.go had to be updated from a mock link
"github.com/tralala/terraform-ansible-provider" to the link of the
actual repository.
Some of the variables in resources code also had to be changed due to not
being descriptive enough, same goes for some of the comments.