Mark vault's yaml property as sensitive
The computed yaml property of the ansible_vault resource should be
marked as sensitive. This content will still be written in plain text to
the state, but we can prevent this from being shown in stdout.

This also makes some minor changes to the test setup to ensure that
things are always cleaned up.