Use a less flakey JSON comparison for tests
The integration tests are a bit flakey with how it compares JSON. The
fixture that's tested against has the Terraform version hard coded in,
so using a different version of Terraform to run tests will cause it to
fail. I've changed it to compare only the relevant portion of the JSON
objects, which helps, but I'm pretty sure the ordering of properties is
not guaranteed so it's possible there will still be intermittent
failures.